chore: clean up and speed up decoding (#10)
* finish adding wasm bindings and asertions

* chore: clean up and speed up decoding

Uses a manual implementation of Deserialize for the graph, such that
nodes and edges are built up gradually during decoding, rather than
decoding the large vectors and then post-processing them. It's a bit
noisy but really not a lot more than what we were doing before, just
wrapped up in serde traits.

Also reworks some code and removes the need for several dependencies.

* rm unused branch